Neurosis 161206/01

+++ controllers/bateria/bateria.php
@@ -34,6 +34,9 @@ class Bateria extends \Libs\Controller {
 
 		$alunos = $this->model->load_active_list('aluno');
 
+		$this->view->set_colunas_datatable(['ID', 'Data de Inicio', 'Data do Fim', 'Ações']);
+		$this->listagem($this->model->load_active_list($this->modulo['modulo']));
+
 		foreach ($alunos as $indice => $aluno) {
 			if($aluno['tipo'] != 1){
 				unset($alunos[$indice]);
@@ -46,8 +49,39 @@ class Bateria extends \Libs\Controller {
 		$this->view->render($this->modulo['modulo'] . '/listagem/listagem');
 	}
 
+	public function listagem($dados_linha){
+		if(empty($dados_linha)){
+			return false;
+		}
+
+		foreach ($dados_linha as $indice => $linha) {
+
+			$data_inicio = date_format(\DateTime::createFromFormat('Y-m-d', $linha['data_inicio']), 'd/m/Y');
+			$data_fim = date_format(\DateTime::createFromFormat('Y-m-d',$linha['data_fim']), 'd/m/Y');
+
+			$retorno_linhas[] = [
+				"<td class='sorting_1'>{$linha['id']}</td>",
+	       		"<td>{$data_inicio}</td>",
+	       		"<td>{$data_fim}</td>",
+	       		"<td>" . $this->view->default_buttons_listagem($linha['id'], false, true, true) . "</td>"
+			];
+		}
+
+		if(!isset($retorno_linhas)){
+			return false;
+		}
+
+		$this->view->linhas_datatable = $retorno_linhas;
+	}
+
 	public function editar($id) {
 		\Util\Permission::check($this->modulo['modulo'], $this->modulo['modulo'] . "_" . "editar");
+
+		if(empty($this->model->db->select("SELECT id FROM {$this->modulo['modulo']} WHERE id = {$id[0]} AND ativo = 1"))){
+			$this->view->alert_js("{$this->modulo['send']} não existe...", 'erro');
+			header('location: ' . URL . $this->modulo['modulo']);
+			exit;
+		}
 
 		$this->view->bateria_list = $this->model->load_active_list($this->modulo['modulo']);
 		$this->view->cadastro = $this->model->full_load_by_id($this->modulo['modulo'], $id[0])[0];
@@ -118,46 +152,79 @@ class Bateria extends \Libs\Controller {
 	public function update($id) {
 		\Util\Permission::check($this->modulo['modulo'], $this->modulo['modulo'] . "_" . "editar");
 
-		$relacao_bateria = $this->model->full_load_by_column('bateria_relaciona_aluno_paciente', 'id_bateria', $id[0]);
-		$update_db = carregar_variavel($this->modulo['modulo']);
-		$relacoes = carregar_variavel('relacao_aluno_paciente');
-
-		debug2($relacao_bateria);
-		debug2($relacoes);
-
-		debug2($update_db);
-				debug2($update_db);
-		exit;
-
-		if($retorno['status']){
-			$relacoes = carregar_variavel('relacao_aluno_paciente');
-
-			foreach ($relacoes as $indice => $relacao) {
-				$insert_ficha_clinica = [
-					'ativo' => 1
-				];
-
-				$retorno_ficha_clinica[$indice] = $this->model->create('ficha_clinica', $insert_ficha_clinica);
-
-				if($retorno_ficha_clinica[$indice]['status']){
-					$insert_relacao = [
-						'id_bateria' 		=> $retorno['id'] ,
-						'id_aluno' 			=> $relacao['relacao']['aluno'],
-						'id_paciente' 		=> $relacao['relacao']['paciente'],
-						'id_ficha_clinica' 	=> $retorno_ficha_clinica[$indice]['id'],
-					];
-
-					$retorno_relacao[$indice] = $this->model->create('bateria_relaciona_aluno_paciente', $insert_relacao);
-				}
-			}
+		if(empty($this->model->db->select("SELECT id FROM {$this->modulo['modulo']} WHERE id = {$id[0]} AND ativo = 1"))){
+			$this->view->alert_js("{$this->modulo['send']} não existe...", 'erro');
+			header('location: ' . URL . $this->modulo['modulo']);
+			exit;
 		}
-
 
 
 		$update_db = carregar_variavel($this->modulo['modulo']);
 		$retorno = $this->model->update($this->modulo['modulo'], $id[0], $update_db);
 
+		if($retorno['status']){
 
+			$relacao_bateria = $this->model->full_load_by_column('bateria_relaciona_aluno_paciente', 'id_bateria', $id[0]);
+			$relacoes = carregar_variavel('relacao_aluno_paciente');
+			unset($relacoes['$id_clone']);
+
+			debug2($relacoes);
+
+			$unset_baterias_existentes = [];
+
+			foreach ($relacao_bateria as $indice => $bateria) {
+				$unset_baterias_existentes[$bateria['id']] = $bateria['id'];
+			}
+
+			$insert_relacao = [];
+
+			foreach ($relacoes as $indice => $relacao) {
+
+				if(empty($relacao['relacao']['id'])){
+					$insert_relacao[] = $relacoes[$indice];
+					unset($relacoes[$indice]);
+				}
+
+				if(isset($relacao['relacao']['id']) && isset($unset_baterias_existentes[$relacao['relacao']['id']]) && $unset_baterias_existentes[$relacao['relacao']['id']] == $relacao['relacao']['id']){
+					unset($relacoes[$indice]);
+					unset($unset_baterias_existentes[$relacao['relacao']['id']]);
+				}
+			}
+
+			if(!empty($insert_relacao)){
+				foreach ($insert_relacao as $indice => $relacao) {
+
+					$insert_ficha_clinica = [
+						'ativo' => 1
+					];
+
+					$retorno_ficha_clinica[$indice] = $this->model->create('ficha_clinica', $insert_ficha_clinica);
+
+					if($retorno_ficha_clinica[$indice]['status']){
+						$insert = [
+							'id_bateria' 		=> $id[0],
+							'id_aluno' 			=> $relacao['relacao']['aluno'],
+							'id_paciente' 		=> $relacao['relacao']['paciente'],
+							'id_ficha_clinica' 	=> $retorno_ficha_clinica[$indice]['id'],
+						];
+
+						$retorno_relacao[$indice] = $this->model->create('bateria_relaciona_aluno_paciente', $insert);
+					}
+				}
+			}
+
+			if(!empty($unset_baterias_existentes)){
+				foreach ($unset_baterias_existentes as $indice => $deletar) {
+
+
+					$update = [
+						'ativo' => 0,
+					];
+
+					$this->model->update('bateria_relaciona_aluno_paciente', $deletar, $update);
+				}
+			}
+		}
 
 		if($retorno['status']){
 			$this->view->alert_js('Cadastro editado com sucesso!!!', 'sucesso');
@@ -170,6 +237,12 @@ class Bateria extends \Libs\Controller {
 
 	public function delete($id) {
 		\Util\Permission::check($this->modulo['modulo'], $this->modulo['modulo'] . "_" . "deletar");
+
+		if(empty($this->model->db->select("SELECT id FROM {$this->modulo['modulo']} WHERE id = {$id[0]} AND ativo = 1"))){
+			$this->view->alert_js("{$this->modulo['send']} não existe...", 'erro');
+			header('location: ' . URL . $this->modulo['modulo']);
+			exit;
+		}
 
 		$retorno = $this->model->delete($this->modulo['modulo'], $id[0]);
