+++ controllers/candidato/candidato.php
@@ -37,7 +37,6 @@ class Candidato extends \Libs\Controller {
 	        $hoje = new \DateTime(date('Y-m-d'));
 	        $diferenca = $nascimento->diff($hoje);
 
-	        // $idade = $diferenca->y . ' anos e ' . $diferenca->m . ' meses';
 	        $idade = $diferenca->y . ' anos';
 
 			$sexo = $linha['sexo'] == 1 ? "Masculino" : "Feminino";
@@ -59,7 +58,7 @@ class Candidato extends \Libs\Controller {
 				"<td>{$idade}</td>",
 				"<td>{$sexo}</td>",
 				"<td>{$linha['patologia']}</td>",
-	        	"<td>" . $this->view->default_buttons_listagem($linha['id']) . $botao_paciente . $botao_ex_paciente . "</td>"
+	        	"<td>" . $this->view->default_buttons_listagem($linha['id'], true, true, false) . $botao_paciente . $botao_ex_paciente . "</td>"
 			];
 		}
 
@@ -70,6 +69,12 @@ class Candidato extends \Libs\Controller {
 		\Util\Permission::check($this->modulo['modulo'], $this->modulo['modulo'] . "_" . "editar");
 
 
+		if(empty($this->model->db->select("SELECT id FROM paciente WHERE id = {$id[0]} AND ativo = 1"))){
+			$this->view->alert_js("{$this->modulo['send']} não existe...", 'erro');
+			header('location: ' . URL . $this->modulo['modulo']);
+			exit;
+		}
+
 		$this->view->cadastro = $this->load_external_model('paciente')->load_paciente($id[0]);
 		$this->view->render($this->modulo['modulo'] . '/editar/editar');
 	}
@@ -77,6 +82,11 @@ class Candidato extends \Libs\Controller {
 	public function visualizar($id){
 		\Util\Permission::check($this->modulo['modulo'], $this->modulo['modulo'] . "_" . "visualizar");
 
+		if(empty($this->model->db->select("SELECT id FROM paciente WHERE id = {$id[0]} AND ativo = 1"))){
+			$this->view->alert_js("{$this->modulo['send']} não existe...", 'erro');
+			header('location: ' . URL . $this->modulo['modulo']);
+			exit;
+		}
 		$this->view->cadastro = $this->load_external_model('paciente')->load_paciente($id[0]);
 		$this->view->render($this->modulo['modulo'] . '/editar/editar');
 
@@ -135,6 +145,12 @@ class Candidato extends \Libs\Controller {
 		\Util\Permission::check($this->modulo['modulo'], $this->modulo['modulo'] . "_" . "editar");
 
 
+		if(empty($this->model->db->select("SELECT id FROM paciente WHERE id = {$id[0]} AND ativo = 1"))){
+			$this->view->alert_js("{$this->modulo['send']} não existe...", 'erro');
+			header('location: ' . URL . $this->modulo['modulo']);
+			exit;
+		}
+
 		$update_db = carregar_variavel($this->modulo['modulo']);
 
 		$update_db += [
@@ -169,6 +185,12 @@ class Candidato extends \Libs\Controller {
 	public function delete($id) {
 		\Util\Permission::check($this->modulo['modulo'], $this->modulo['modulo'] . "_" . "deletar");
 
+		if(empty($this->model->db->select("SELECT id FROM paciente WHERE id = {$id[0]} AND ativo = 1"))){
+			$this->view->alert_js("{$this->modulo['send']} não existe...", 'erro');
+			header('location: ' . URL . $this->modulo['modulo']);
+			exit;
+		}
+
 		$retorno_paciente = $this->model->delete('paciente', $id[0]);
 
 		if($retorno_paciente['status']){
@@ -187,6 +209,14 @@ class Candidato extends \Libs\Controller {
 
 	public function transformar_paciente($id){
 
+		\Util\Permission::check($this->modulo['modulo'], $this->modulo['modulo'] . "_" . "transformar_paciente");
+
+		if(empty($this->model->db->select("SELECT id FROM paciente WHERE id = {$id[0]} AND ativo = 1"))){
+			$this->view->alert_js("{$this->modulo['send']} não existe...", 'erro');
+			header('location: ' . URL . $this->modulo['modulo']);
+			exit;
+		}
+
 		$update_db = [
 			"tipo" => 1
 		];
@@ -204,6 +234,14 @@ class Candidato extends \Libs\Controller {
 
 	public function transformar_ex_paciente($id){
 
+		\Util\Permission::check($this->modulo['modulo'], $this->modulo['modulo'] . "_" . "transformar_ex_paciente");
+
+		if(empty($this->model->db->select("SELECT id FROM paciente WHERE id = {$id[0]} AND ativo = 1"))){
+			$this->view->alert_js("{$this->modulo['send']} não existe...", 'erro');
+			header('location: ' . URL . $this->modulo['modulo']);
+			exit;
+		}
+
 		$update_db = [
 			"tipo" => 2
 		];
@@ -220,6 +258,15 @@ class Candidato extends \Libs\Controller {
 	}
 
 	public function transformar_candidato($id){
+
+		\Util\Permission::check($this->modulo['modulo'], $this->modulo['modulo'] . "_" . "transformar_candidato");
+
+
+		if(empty($this->model->db->select("SELECT id FROM paciente WHERE id = {$id[0]} AND ativo = 1"))){
+			$this->view->alert_js("{$this->modulo['send']} não existe...", 'erro');
+			header('location: ' . URL . $this->modulo['modulo']);
+			exit;
+		}
 
 		$update_db = [
 			"tipo" => 0
